requirements.txt
numpy~=2.2.6
pillow~=11.3.0
h11~=0.16.0
pip~=24.0
attrs~=25.3.0
distro~=1.9.0
cryptography~=45.0.7
Jinja2~=3.1.6
filelock~=3.19.1
redis~=5.2.1
rsa~=4.9.1
pyasn1~=0.6.1
dill~=0.4.0
Mako~=1.3.10
Pygments~=2.19.2
MarkupSafe~=3.0.2
pytz~=2025.2
rich~=14.1.0
markdown-it-py~=4.0.0
requests~=2.32.5
colorama~=0.4.6
PyYAML~=6.0.2
idna~=3.10
sniffio~=1.3.1
click~=8.2.1
httpcore~=1.0.9
httpx~=0.28.1
zstandard~=0.24.0
jiter~=0.10.0
mdurl~=0.1.2
charset-normalizer~=3.4.3
setuptools~=80.9.0
cffi~=1.17.1
langchain-core~=0.3.75
langchain-openai~=0.3.32
pydantic~=2.11.7
typer~=0.16.1
tqdm~=4.67.1
langchain~=0.3.27
tiktoken~=0.11.0
pandas~=2.3.2
langsmith~=0.4.24
tenacity~=9.1.2
regex~=2025.9.1
scipy~=1.16.1
threadpoolctl~=3.6.0
torch~=2.8.0
mpmath~=1.3.0
sympy~=1.14.0
packaging~=25.0
limits~=5.5.0
shellingham~=1.5.4
fsspec~=2025.9.0
certifi~=2025.8.3
urllib3~=2.5.0
openai~=1.106.1
anyio~=4.10.0
orjson~=3.11.3
openpyxl~=3.1.5
SQLAlchemy~=2.0.43
lxml~=5.4.0
python-dateutil~=2.9.0.post0
xxhash~=3.5.0
python-multipart~=0.0.20
alembic~=1.16.5
six~=1.17.0
tzdata~=2025.2
h2~=4.3.0
uvicorn~=0.35.0
bcrypt~=5.0.0
fastapi~=0.116.1
jsonschema~=4.25.1
psutil~=7.0.0
starlette~=0.47.3
huggingface-hub~=0.34.4
multiprocess~=0.70.18
filetype~=1.2.0
greenlet~=3.2.4
networkx~=3.5
email-validator~=2.3.0
Deprecated~=1.2.18
annotated-types~=0.7.0
langgraph~=0.6.7
ormsgpack~=1.10.0
requests-toolbelt~=1.0.0
groq~=0.31.1
jsonref~=1.1.0
referencing~=0.36.2
jsonpointer~=3.0.0
jsonschema-specifications~=2025.4.1
psycopg2~=2.9.10
tokenizers~=0.22.0
transformers~=4.56.0
jsonpatch~=1.33
langchain-text-splitters~=0.3.11
safetensors~=0.6.2
pydantic-settings~=2.10.1
python-dotenv~=1.1.1
pypdfium2~=4.30.0
beautifulsoup4~=4.13.5
sentence-transformers~=5.1.0